refactor(Editor): Took Redux logic out of Editor view, made Codemirror a HOC
style(editor): Fix linting errors
